requirements.txt
# Webセキュリティ診断自動化システム 依存関係

# Web関連
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0

# ブラウザ自動化
selenium>=4.15.0

# データ処理
pandas>=2.1.0

# 設定・テンプレート
pyyaml>=6.0.1
jinja2>=3.1.2
python-dotenv>=1.0.0

# ログ・監視
structlog>=23.1.0
prometheus-client>=0.17.0

# セキュリティ
cryptography>=41.0.0

# キャッシュ
redis>=4.6.0

# テスト
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# 開発・デバッグ
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
